requirements-dev.txt
-r requirements.txt
betamax==0.4.2
mrproxy==0.3.3
py==1.4.26
pytest==2.7.0
pytest-django==2.8.0
pytest-pythonpath==0.6
Pygments==2.0.1
Sphinx==1.2.3
argh==0.26.1
docutils==0.12
pathtools==0.1.2
requests==2.7.0
sphinx-rtd-theme==0.1.6
